pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PokeELT"
version = "0.1.0"
dependencies = [
  "duckdb>=1.2.0",
  "requests==2.32.3",
  "PyYAML==6.0.2",
]
requires-python = ">=3.10"
authors = [{name = "Jason Anderson", email = "jasreidan@gmail.com"}]
maintainers = [{name = "Jason Anderson", email = "jasreidan@gmail.com"}]
description = "Extracts data about Pokemon from a public API and stores it in a duckdb database. Future plans include modeling in dbt and dashboarding in Streamlit."
readme = "README.md"
license = "Apache-2.0"

[project.urls]
Repository = "https://github.com/Jace743/PokeELT.git"